enable eu logger
