+++ src/promptflow/promptflow/_sdk/_configuration.py
@@ -28,6 +28,8 @@ class Configuration(object):
     CONFIG_PATH = Path.home() / ".promptflow" / "pf.yaml"
     COLLECT_TELEMETRY = "cli.telemetry_enabled"
     EXTENSION_COLLECT_TELEMETRY = "extension.telemetry_enabled"
+    EU_USER = "cli.eu_user"
+    EXTENSION_EU_USER = "extension.eu_user"
     INSTALLATION_ID = "cli.installation_id"
     CONNECTION_PROVIDER = "connection.provider"
     _instance = None
@@ -154,6 +156,12 @@ class Configuration(object):
     def set_telemetry_consent(self, value):
         """Set the telemetry consent value and store in local."""
         self.set_config(key=self.COLLECT_TELEMETRY, value=value)
+
+    def is_eu_user(self) -> Optional[bool]:
+        """Check if user is from europe. Return None if not configured."""
+        if call_from_extension():
+            return self.get_config(key=self.EXTENSION_EU_USER)
+        return self.get_config(key=self.EU_USER)
 
     def get_or_set_installation_id(self):
         """Get user id if exists, otherwise set installation id and return it."""
