+++ src/promptflow/tests/sdk_cli_azure_test/conftest.py
@@ -7,6 +7,7 @@ from azure.ai.ml import MLClient
 from azure.ai.ml.entities import Data
 from azure.core.exceptions import ResourceNotFoundError
 
+from promptflow._utils.utils import environment_variable_overwrite
 from promptflow.azure import PFClient
 
 from ._azure_utils import get_cred
@@ -34,28 +35,32 @@ def ml_client(
 
 @pytest.fixture()
 def remote_client() -> PFClient:
-    return PFClient(
-        credential=get_cred(),
-        subscription_id="96aede12-2f73-41cb-b983-6d11a904839b",
-        resource_group_name="promptflow",
-        workspace_name="promptflow-eastus",
-    )
+    # enable telemetry for CI
+    with environment_variable_overwrite("TELEMETRY_ENABLED", "true"):
+        yield PFClient(
+            credential=get_cred(),
+            subscription_id="96aede12-2f73-41cb-b983-6d11a904839b",
+            resource_group_name="promptflow",
+            workspace_name="promptflow-eastus",
+        )
 
 
 @pytest.fixture()
 def remote_client_int() -> PFClient:
-    client = MLClient(
-        credential=get_cred(),
-        subscription_id="96aede12-2f73-41cb-b983-6d11a904839b",
-        resource_group_name="promptflow",
-        workspace_name="promptflow-int",
-    )
-    return PFClient(ml_client=client)
+    # enable telemetry for CI
+    with environment_variable_overwrite("TELEMETRY_ENABLED", "true"):
+        client = MLClient(
+            credential=get_cred(),
+            subscription_id="96aede12-2f73-41cb-b983-6d11a904839b",
+            resource_group_name="promptflow",
+            workspace_name="promptflow-int",
+        )
+        yield PFClient(ml_client=client)
 
 
 @pytest.fixture()
 def pf(remote_client) -> PFClient:
-    return remote_client
+    yield remote_client
 
 
 @pytest.fixture
