+++ src/promptflow/tests/sdk_cli_test/conftest.py
@@ -9,6 +9,7 @@ from promptflow import PFClient
 from promptflow._sdk._serving.app import create_app as create_serving_app
 from promptflow._sdk.entities import AzureOpenAIConnection as AzureOpenAIConnectionEntity
 from promptflow._sdk.entities._connection import CustomConnection, _Connection
+from promptflow._utils.utils import environment_variable_overwrite
 
 PROMOTFLOW_ROOT = Path(__file__) / "../../.."
 RUNTIME_TEST_CONFIGS_ROOT = Path(PROMOTFLOW_ROOT / "tests/test_configs/runtime")
@@ -18,12 +19,16 @@ MODEL_ROOT = Path(PROMOTFLOW_ROOT / "tests/test_configs/flows")
 
 @pytest.fixture(scope="session")
 def local_client() -> PFClient:
-    return PFClient()
+    # enable telemetry for CI
+    with environment_variable_overwrite("TELEMETRY_ENABLED", "true"):
+        yield PFClient()
 
 
 @pytest.fixture(scope="session")
 def pf() -> PFClient:
-    return PFClient()
+    # enable telemetry for CI
+    with environment_variable_overwrite("TELEMETRY_ENABLED", "true"):
+        return PFClient()
 
 
 @pytest.fixture()
